dtc_de/requirements.extract_load_trips_from_tlc_to_gs.txt
aiofiles==23.1.0
aiohttp==3.8.4
google-cloud-storage==2.8.0
